[Refactor] Replace direct type serialization and deserialization with TypeSerializer and TypeDeserializer
