+++ fe/fe-core/src/main/java/com/starrocks/catalog/combinator/AggStateDesc.java
@@ -22,6 +22,7 @@ import com.starrocks.catalog.AggregateFunction;
 import com.starrocks.catalog.Function;
 import com.starrocks.catalog.FunctionSet;
 import com.starrocks.catalog.Type;
+import com.starrocks.catalog.TypeSerializer;
 import com.starrocks.common.AnalysisException;
 import com.starrocks.qe.ConnectContext;
 import com.starrocks.sql.analyzer.FunctionAnalyzer;
@@ -147,7 +148,7 @@ public class AggStateDesc {
         for (Type argType : argTypes) {
             TTypeDesc tTypeDesc = new TTypeDesc();
             tTypeDesc.setTypes(new ArrayList<TTypeNode>());
-            argType.toThrift(tTypeDesc);
+            TypeSerializer.toThrift(argType, tTypeDesc);
             tAggStateDesc.addToArg_types(tTypeDesc);
         }
         tAggStateDesc.setResult_nullable(resultNullable);
@@ -156,7 +157,7 @@ public class AggStateDesc {
         // ret type
         TTypeDesc tTypeDesc = new TTypeDesc();
         tTypeDesc.setTypes(new ArrayList<TTypeNode>());
-        returnType.toThrift(tTypeDesc);
+        TypeSerializer.toThrift(returnType, tTypeDesc);
         tAggStateDesc.setRet_type(tTypeDesc);
         Preconditions.checkState(!tTypeDesc.types.isEmpty());
